Updating code to PuTTY 25b034ee
Source commit: 3a37adcc642933a71ce6554a76462bf55ceb5a43

+++ source/putty/misc.h
@@ -297,7 +297,7 @@ void debug_memdump(const void *buf, int len, bool L);
 #ifdef MPEXT
 // Recent PuTTY code uses C99 standard that allows code before initialization.
 // Mostly that code are assertions. This assert implementation allows being used before code.
-#define pinitassert(P) const int __assert_dummy = 1/(P)
+#define pinitassert(P) const int __assert_dummy = 1/((int)(P))
 #endif
 
 #endif
